Serperate app from server
