+++ package.json
@@ -4,7 +4,8 @@
   "description": "A simple todolist meant for learning Javascript.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "server": "nodemon --exec babel-node server/index.js"
   },
   "repository": {
     "type": "git",
@@ -16,5 +17,12 @@
   "bugs": {
     "url": "https://github.com/ojjo-sh/toodledoodledoo/issues"
   },
-  "homepage": "https://github.com/ojjo-sh/toodledoodledoo#readme"
+  "homepage": "https://github.com/ojjo-sh/toodledoodledoo#readme",
+  "devDependencies": {
+    "@babel/cli": "^7.7.7",
+    "@babel/core": "^7.7.7",
+    "@babel/node": "^7.7.7",
+    "@babel/preset-env": "^7.7.7",
+    "nodemon": "^1.19.4"
+  }
 }
