+++ server/utils/index.js
@@ -1,8 +1,7 @@
 import fs from 'fs';
 
-export function doesFileExist(filePath) {
+export function doesFileExist (filePath) {
     try {
-
         if (typeof filePath !== 'string') {
             throw new Error("Error: Filepath is not of type string");
         }
@@ -12,7 +11,6 @@ export function doesFileExist(filePath) {
         } else {
             return false;
         }
-
     } catch (err) {
         console.error(err)
     }
